Add shortcut to API
